Add initial tests for endpoint behavior
Create endpointBehavior() function in test.js. This function is a
standardized set of assertions that all endpoints in the API should
pass.

Add two initial tests to the function: one that checks for hypermedia
type application/hal+json in the response body (if the client accepts it),
and another that checks if the endpoint responds with 406 if the client
only accepts a nonsense media type.

Addtionally, add .eslintrc.json to the commit to standardize linting
across the project (primarily using the Airbnb styleguide eslint
plugin).