Add error handling and tests for bad HTTP methods
Add an all route to server that checks for allowed HTTP methods on
a resource (currently hardcoded to just GET on the API root). If the
client uses a different verb, route to the error-handling middleware
at the end of the server code (which will return 405 and customized
error in response to the client).

Add endpointMethods() function, which takes an endpoint and a list
of methods and asserts that all other common HTTP methods not in that
list should get a 405 response.