Add all models
In addition to the User model, add Message and Log models. Messages
are written by a single user and have a text property. Logs have a
list of users (as log can be between just one user or many) and a list
of messages.

(Also, lint syntax across the project).