Add express default error handler
Add 'custom' property to all thrown errors, and check for it in the
error handler. If the property is present, the handler with generate
and represent the custom message for the client. Otherwise, the default
handler will take over.

Additionally, if the response has been partially constructed,
the default error handler will take over.