Add POST to /api/v1/users
Install body-parser module, and include it as a middleware function
to parse POST request body. If the JSON in the body is properly formatted,
then create a new user in the database and serve a response to the client
with the location of the new resource.

Write a generic POST test that checks for a 201 response.