Add PUT routing for /api/v1/users/:name
Multiple middleware functions handle different parts of the PUT
request--the first checks if the request is an object, the second
checks for the presense of a 'name' key in the request, and updates
the database if so, and the last route renders a response.

Write generic PUT tests--idempotency (importantly, not used for this
first PUT route, as changing the name of the user changes the resource
representation location), 200 response, invalide media types.