Add PUT routing for /api/v1/logs/:name
Multiple middleware functions handle different parts of the PUT
request:

- The first checks if the request is an objects
- The second checks for the presense of a 'users' key in the request,
makes sure they're existing user names, and updates the database if so
- The third checks for the presense of a 'messages' key in the request,
makes sure they're existing messages, and updates the database if so
- The third checks for the presense of a 'name' key in the request,
and updates the database if so
- The last route renders a response

Apply generic PUT tests to the route, including idempotency. Update
generic idempotency test to have second request follow the Location
header in the response to the first request.